Refactor out the property setup for task and extension onto traits
